configs: add defconfig for TS-4800
This patch adds support for the TS-4800 SoM manufactured by Technologic
Systems based on the Freescale i.MX515 SoC.

The defconfig uses a custom Linux based on v4.3. Hopefully, the TS-4800
support will be merged mainline at some point, and it will be possible
to use a vanilla kernel.

There are also two patches for U-Boot. One of them adds support for
device trees and is already upstream but not yet in master. There other
one modifies the default environment to more sane values and should be
sent for upstream soon.

Finally, mbrboot.bin is a MBR routine which was extracted from the image
provided by Technologic Systems. Basically, it scans every partition
with the 0xda type and copies the first file to a given offset, and then
jump where the first file was copied to start its execution. In our
case, we use it to load u-boot.

A post-image script is provided to generate an image that can be "dd"
directly to an sdcard.

More details on the board here:
  http://wiki.embeddedarm.com/wiki/TS-4800

Signed-off-by: Damien Riegel <damien.riegel@savoirfairelinux.com>